+++ test/studio.test.ts
@@ -1309,6 +1309,7 @@ describe('Studio API: home page', () => {
       expect(catalog[page]).not.toContain('main-search')
       expect(catalog[page]).not.toContain('main-blog')
       expect(catalog[page]).not.toContain('main-article')
+      expect(catalog[page]).not.toContain('main-404')
       expect(catalog[page]).not.toContain('predictive-search')
       expect(catalog[page]).not.toContain('quick-add')
     }